+++ modules/loadout/class-loadout-module.php
@@ -102,6 +102,7 @@ class Loadout_Module extends FFLA_Module
                     \Bricks\Elements::register_element($path . 'frontend/class-loadout-add-button-element.php');
                     \Bricks\Elements::register_element($path . 'frontend/class-loadout-add-tier-element.php');
                     \Bricks\Elements::register_element($path . 'frontend/class-loadout-tier-tabs-element.php');
+                    \Bricks\Elements::register_element($path . 'frontend/class-loadout-products-element.php');
                     \Bricks\Elements::register_element($path . 'frontend/class-loadout-progress-element.php');
                     \Bricks\Elements::register_element($path . 'frontend/class-loadout-cart-mirror-element.php');
                 }